requirements_mcp.txt
# ScreenCoder MCP 工具依赖

# MCP 核心依赖
mcp>=1.0.0

# 原有 ScreenCoder 依赖
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
scikit-learn>=1.3.0
scipy>=1.11.0

# AI 模型客户端
openai>=1.0.0
google-generativeai>=0.3.0
dashscope>=1.14.0  # 阿里云Qwen
volcengine>=1.0.0  # 字节跳动Doubao

# 异步和并发
asyncio-throttle>=1.0.0
aiofiles>=23.0.0

# 配置和日志
pydantic>=2.0.0
loguru>=0.7.0

# 图像处理
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Web 相关
requests>=2.31.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.0.0
